% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{clone_assignments}
\alias{clone_assignments}
\title{Clone assignment repositories from GitHub Classroom}
\usage{
clone_assignments(path, assignment, organisation)
}
\arguments{
\item{path}{The directory to clone repositories into. Each repository will be
added in a subdirectory based on the student's GitHub username.}

\item{assignment}{The assignment to clone (the prefix of the assignment
repository name on GitHub).}

\item{organisation}{The GitHub Classroom organisation name.}
}
\description{
\lifecycle{maturing}
}
\examples{

\dontrun{
clone_assignments(
  "~/teaching/ETC5523/assignment_1",
  assignment = "take-home-assignment",
  organisation = "etc5523-2020"
 )
}

}
